Moved benchmarking/utilities to source/tomopy/misc/benchmark.py
- Moving to this location is breaking pyctest
