+++ source/tomopy/misc/benchmark.py
